requirements.txt
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.3.0
pyscreenshot==3.0
requests==2.25.1
torch==1.13.0
torchvision==0.14.0
